+++ CHANGELOG.md
@@ -12,11 +12,14 @@ This project uses [CalVer](https://calver.org/) (YY.MM.Micro) for product versio
 - Retry failed evaluation extraction runs from the UI
 - Require users to accept General Conditions of Service and Privacy Policy before accessing the app
 - Manage General Conditions of Service and Privacy Policy from the Backoffice
+- (beta) Medgemma 1.0 available for all Agents
+- (beta) Gemma 4 available for all Agents
 
 ### Changed
 
 ### Fixed
 - Recover stuck evaluation extraction worker jobs after rebooting
+- Langfuse trace
 
 ### Security